haku/console: split operator OAuth + server catalog out of mcp_approval
mcp_approval.py had grown well beyond tool-call approval — the connected
MCP-server catalog/credentials and the whole operator-OAuth account linkage.
Reorganize into three domain modules, with several cleanups along the way.

Modules:
- mcp_config.py: the connected-MCP-server catalog, transport selection, and
  static bearer credentials — the substrate both routers share.
- mcp_operator_oauth.py: the DCR/PKCE flow, Postgres token storage/refresh, and
  the /api/mcp/operator-auth/* connect/disconnect/callback endpoints.
- mcp_approval.py: now just the tool-call ledger, execution, event fan-out,
  tool reflection, and approval endpoints.

Cleanups:
- Replace the dict[str, Any] `_oauth_*_snapshot` helpers with typed Pydantic
  models (OperatorOAuthFlowState / OperatorOAuthRefreshState) that carry detached
  row state across the token exchange/refresh network calls.
- Move the OAuth callback page out of a hardcoded Python f-string into a sibling
  mcp_operator_auth_callback.html template (string.Template); TODO to make it a
  SPA-style page.
- Drop the single-implementation McpOperatorOAuthStoreProtocol and
  ToolCallLedgerProtocol — one Postgres backend each, tests run against real
  Postgres — so DI uses the concrete classes.
- Standardize on `import datetime` / `datetime.datetime` (matching the schema and
  shared tool_calls models).
- Make McpOperatorAuthStatus a `status`-discriminated union
  (McpOperatorAuthConnected | McpOperatorAuthUnconnected) so connected-only fields
  can't appear on an unconnected status.
- Extract _resolve_operator_oauth_client() (metadata discovery + client
  registration I/O) out of _build_operator_oauth_flow().

The wire contract (endpoints, response models) is otherwise unchanged; all callers
are updated atomically with no compatibility shims. Also notes in haku/TODO.md that
launch-routine could fold into the same MCP tool-call approval mechanism.
