+++ haku/console/mcp_config.py
@@ -1,0 +1,103 @@
+"""The connected-MCP-server catalog and how to reach each entry.
+
+The console's deploy-time YAML names the MCP servers Haku may drive through the approval
+queue; this module models that config, looks entries up by id, and resolves how to reach
+each one — the in-process `FastMCP` transport or remote URL, and the static bearer
+credential where one applies. Both the approval router (`mcp_approval`) and the operator
+OAuth linkage (`mcp_operator_oauth`) build on this shared substrate.
+"""
+
+from __future__ import annotations
+
+import os
+import re
+
+import yaml
+from fastapi import HTTPException
+from fastmcp import FastMCP
+from pydantic import BaseModel, Field
+
+from haku.console.config import Settings
+
+
+class McpOperatorOAuthConfig(BaseModel):
+    enabled: bool = True
+    client_name: str = "Haku Console"
+    scopes: list[str] | None = None
+    # For an authorization server with no open Dynamic Client Registration (RFC 7591) —
+    # e.g. Authentik, which has no /register endpoint, so the server-metadata-declared
+    # registration_endpoint is absent and DCR would otherwise 401 against a guessed
+    # {server}/register fallback. A pre-registered public/PKCE client_id shared across
+    # every OAuth caller of that authorization server, skipping registration entirely.
+    # Safe to share: PKCE plus per-request redirect_uri validation secure each caller's
+    # auth code exchange independently even though the client_id is the same for all.
+    static_client_id: str | None = None
+
+
+class McpServerEntry(BaseModel):
+    id: str
+    # None for a server reached via an in-process FastMCP instance instead of a remote
+    # URL (see McpToolExecutor/McpMetadataProvider's `in_process_servers` registry,
+    # e.g. haku.console.tools.gmail's `gmail` server) — resolved at runtime by id,
+    # not by anything in this config model.
+    server_url: str | None = None
+    bearer_token_secret: str | None = None
+    operator_oauth: McpOperatorOAuthConfig | None = None
+
+
+class ConsoleMcpConfig(BaseModel):
+    servers: list[McpServerEntry] = Field(default_factory=list)
+
+
+class ConsoleConfigFile(BaseModel):
+    mcp: ConsoleMcpConfig = Field(default_factory=ConsoleMcpConfig)
+
+
+def _load_servers(settings: Settings) -> list[McpServerEntry]:
+    path = settings.config_file
+    if path is None or not path.exists():
+        return []
+    raw = yaml.safe_load(path.read_text(encoding="utf-8")) or {}
+    return ConsoleConfigFile.model_validate(raw).mcp.servers
+
+
+def _server_entry(settings: Settings, server_id: str) -> McpServerEntry:
+    for server in _load_servers(settings):
+        if server.id == server_id:
+            return server
+    raise HTTPException(status_code=404, detail=f"unknown MCP server: {server_id}")
+
+
+def _operator_oauth_enabled(server: McpServerEntry) -> bool:
+    return bool(server.operator_oauth and server.operator_oauth.enabled)
+
+
+def _credential_env_name(bearer_token_secret: str) -> str:
+    suffix = re.sub(r"[^A-Za-z0-9]+", "_", bearer_token_secret).strip("_").upper()
+    return f"HAKU_CONSOLE_MCP_CREDENTIAL_{suffix}"
+
+
+def _credential_token(server: McpServerEntry) -> str | None:
+    if server.bearer_token_secret is None:
+        return None
+    env_name = _credential_env_name(server.bearer_token_secret)
+    token = os.environ.get(env_name)
+    if not token:
+        raise RuntimeError(f"missing MCP bearer token env var {env_name} for MCP server {server.id}")
+    return token
+
+
+# A server reached over an in-process FastMCP instance instead of a remote URL (see
+# McpServerEntry.server_url). `fastmcp.client.Client` accepts a `FastMCP` instance
+# directly and opens an in-memory `FastMCPTransport` — so both `McpToolExecutor` and
+# `McpMetadataProvider` run the exact same `Client(...)` calls either way; only this
+# lookup differs.
+InProcessServers = dict[str, FastMCP]
+
+
+def _transport(server: McpServerEntry, in_process: InProcessServers) -> FastMCP | str:
+    if in_process_server := in_process.get(server.id):
+        return in_process_server
+    if server.server_url is not None:
+        return server.server_url
+    raise RuntimeError(f"MCP server {server.id!r} has no server_url and no in-process registration")
