+++ haku/console/mcp_operator_auth_callback.html
@@ -1,0 +1,32 @@
+<!doctype html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <title>$title</title>
+    <style>
+      body {
+        color-scheme: light dark;
+        font:
+          16px system-ui,
+          sans-serif;
+        margin: 3rem auto;
+        max-width: 36rem;
+        line-height: 1.4;
+      }
+      code {
+        overflow-wrap: anywhere;
+      }
+    </style>
+  </head>
+  <body>
+    <h1>$title</h1>
+    <p><code>$message</code></p>
+    <!-- Notify any open console tab (the Settings page listens on this channel) that the
+         operator's MCP account linkage changed, so it refetches without a manual reload. -->
+    <script>
+      try {
+        new BroadcastChannel("haku-console-mcp-auth").postMessage({ type: "$payload" });
+      } catch (_) {}
+    </script>
+  </body>
+</html>
