+++ haku/console/mcp_operator_oauth.py
@@ -1,0 +1,661 @@
+"""Operator OAuth account linkage for connected MCP servers.
+
+Some MCP servers execute a tool call as *the operator's own account* rather than under a
+static console-held bearer (e.g. `kubectl-passthrough-mcp`, which runs kubectl as the
+approving operator's cluster-admin identity). For those, the operator links their account
+once through an OAuth authorization-code + PKCE flow; the console runs Dynamic Client
+Registration (or uses a pre-registered `static_client_id`), stores the resulting token
+association, and refreshes it as needed. This module owns that flow, its Postgres-backed
+storage, and the connect/disconnect/callback endpoints. The approval router
+(`mcp_approval`) consumes the linked token via `access_token_for` when it executes an
+approved call; the catalog of which servers use operator OAuth lives in `mcp_config`.
+"""
+
+from __future__ import annotations
+
+import base64
+import datetime
+import html
+import secrets
+from pathlib import Path
+from string import Template
+from typing import Annotated, Literal, cast
+from urllib.parse import quote, urlencode, urljoin
+
+import httpx
+from fastapi import APIRouter, Depends, HTTPException, Request
+from fastapi.responses import HTMLResponse
+from fastapi_csrf_protect import CsrfProtect
+from mcp.client.auth.oauth2 import PKCEParameters
+from mcp.client.auth.utils import (
+    build_oauth_authorization_server_metadata_discovery_urls,
+    build_protected_resource_metadata_discovery_urls,
+    extract_resource_metadata_from_www_auth,
+    extract_scope_from_www_auth,
+    get_client_metadata_scopes,
+    handle_auth_metadata_response,
+    handle_protected_resource_response,
+    handle_registration_response,
+    handle_token_response_scopes,
+)
+from mcp.client.streamable_http import MCP_PROTOCOL_VERSION
+from mcp.shared.auth import (
+    OAuthClientInformationFull,
+    OAuthClientMetadata,
+    OAuthMetadata,
+    OAuthToken,
+    ProtectedResourceMetadata,
+)
+from mcp.shared.auth_utils import check_resource_allowed, resource_url_from_server_url
+from mcp.types import LATEST_PROTOCOL_VERSION
+from pydantic import BaseModel, Field, ValidationError
+from sqlalchemy import create_engine, delete, select
+from sqlalchemy.orm import sessionmaker
+
+from haku.console.config import Settings
+from haku.console.database_migrate import run_migrations_for_connection
+from haku.console.database_schema import McpOperatorOAuthAssociation, McpOperatorOAuthFlow
+from haku.console.deps import SettingsDep
+from haku.console.mcp_config import McpServerEntry, _load_servers, _operator_oauth_enabled, _server_entry
+
+Csrf = Annotated[CsrfProtect, Depends()]
+
+MCP_OPERATOR_AUTH_CALLBACK_PATH = "/api/mcp/operator-auth/callback"
+MCP_OPERATOR_AUTH_FLOW_TTL = datetime.timedelta(minutes=10)
+MCP_OPERATOR_AUTH_REFRESH_SKEW = datetime.timedelta(seconds=60)
+
+router = APIRouter(tags=["mcp-operator-oauth"])
+
+
+class McpOperatorAuthStatusBase(BaseModel):
+    server_id: str
+    operator_principal: str
+
+
+class McpOperatorAuthConnected(McpOperatorAuthStatusBase):
+    status: Literal["connected"] = "connected"
+    connected_at: datetime.datetime
+    # None when the linked token declares no expiry (OAuth `expires_in` absent).
+    token_expires_at: datetime.datetime | None = None
+    scope: str | None = None
+
+
+class McpOperatorAuthUnconnected(McpOperatorAuthStatusBase):
+    status: Literal["unconnected"] = "unconnected"
+
+
+# Discriminated on `status`, so the connected-only fields (connected_at/token_expires_at/
+# scope) exist exactly when connected — no "unconnected with a connected_at" nonsense state.
+type McpOperatorAuthStatus = Annotated[
+    McpOperatorAuthConnected | McpOperatorAuthUnconnected, Field(discriminator="status")
+]
+
+
+class McpOperatorAuthStatusResponse(BaseModel):
+    associations: list[McpOperatorAuthStatus] = Field(default_factory=list)
+
+
+class McpOperatorAuthStartResponse(BaseModel):
+    server_id: str
+    authorization_url: str
+    expires_at: datetime.datetime
+
+
+class _OperatorOAuthTokenClient(BaseModel):
+    """Token-endpoint call parameters shared by the auth-code exchange and the refresh.
+    Detached from the ORM row so the network call runs with no DB session held open."""
+
+    client_id: str
+    client_secret: str | None = None
+    token_endpoint_auth_method: str | None = None
+    token_endpoint: str
+    resource: str | None = None
+
+
+class OperatorOAuthFlowState(_OperatorOAuthTokenClient):
+    """An `McpOperatorOAuthFlow` row read out before its session closes, carried across the
+    authorization-code → token exchange (which must not hold a DB session open)."""
+
+    server_id: str
+    operator_principal: str
+    expires_at: datetime.datetime
+    redirect_uri: str
+    code_verifier: str
+    client_secret_expires_at: int | None = None
+    scope: str | None = None
+
+    @classmethod
+    def from_row(cls, row: McpOperatorOAuthFlow) -> OperatorOAuthFlowState:
+        return cls(
+            client_id=row.client_id,
+            client_secret=row.client_secret,
+            token_endpoint_auth_method=row.token_endpoint_auth_method,
+            token_endpoint=row.token_endpoint,
+            resource=row.resource,
+            server_id=row.server_id,
+            operator_principal=row.operator_principal,
+            expires_at=row.expires_at,
+            redirect_uri=row.redirect_uri,
+            code_verifier=row.code_verifier,
+            client_secret_expires_at=row.client_secret_expires_at,
+            scope=row.scope,
+        )
+
+
+class OperatorOAuthRefreshState(_OperatorOAuthTokenClient):
+    """An `McpOperatorOAuthAssociation` row's refresh inputs read out before its session
+    closes, carried across the token-refresh network call."""
+
+    refresh_token: str | None = None
+
+    @classmethod
+    def from_row(cls, row: McpOperatorOAuthAssociation) -> OperatorOAuthRefreshState:
+        return cls(
+            client_id=row.client_id,
+            client_secret=row.client_secret,
+            token_endpoint_auth_method=row.token_endpoint_auth_method,
+            token_endpoint=row.token_endpoint,
+            resource=row.resource,
+            refresh_token=row.refresh_token,
+        )
+
+
+class _BuiltOperatorOAuthFlow(BaseModel):
+    state: str
+    authorization_url: str
+    expires_at: datetime.datetime
+    redirect_uri: str
+    code_verifier: str
+    client_info: OAuthClientInformationFull
+    token_endpoint: str
+    resource: str | None = None
+    scope: str | None = None
+
+
+class PostgresMcpOperatorOAuthStore:
+    """Postgres-backed operator OAuth association store for connected MCP servers."""
+
+    def __init__(self, database_url: str) -> None:
+        self._engine = create_engine(database_url, pool_pre_ping=True)
+        with self._engine.begin() as conn:
+            run_migrations_for_connection(conn)
+        self._sessions = sessionmaker(self._engine, expire_on_commit=False)
+
+    def list_statuses(self, *, servers: list[McpServerEntry], operator_principal: str) -> McpOperatorAuthStatusResponse:
+        oauth_servers = [server for server in servers if _operator_oauth_enabled(server)]
+        if not oauth_servers:
+            return McpOperatorAuthStatusResponse()
+        server_ids = [server.id for server in oauth_servers]
+        with self._sessions.begin() as session:
+            rows = session.scalars(
+                select(McpOperatorOAuthAssociation)
+                .where(McpOperatorOAuthAssociation.operator_principal == operator_principal)
+                .where(McpOperatorOAuthAssociation.server_id.in_(server_ids))
+            ).all()
+        by_server = {row.server_id: row for row in rows}
+        return McpOperatorAuthStatusResponse(
+            associations=[
+                _oauth_status_from_row(server.id, operator_principal, by_server.get(server.id))
+                for server in oauth_servers
+            ]
+        )
+
+    async def start_flow(
+        self, *, server: McpServerEntry, operator_principal: str, public_base_url: str
+    ) -> McpOperatorAuthStartResponse:
+        if not _operator_oauth_enabled(server):
+            raise HTTPException(status_code=404, detail=f"MCP server {server.id} does not use operator OAuth")
+        flow = await _build_operator_oauth_flow(server, public_base_url.rstrip("/"))
+        with self._sessions.begin() as session:
+            now = datetime.datetime.now(datetime.UTC)
+            session.execute(delete(McpOperatorOAuthFlow).where(McpOperatorOAuthFlow.expires_at < now))
+            session.execute(
+                delete(McpOperatorOAuthFlow)
+                .where(McpOperatorOAuthFlow.server_id == server.id)
+                .where(McpOperatorOAuthFlow.operator_principal == operator_principal)
+            )
+            session.add(
+                McpOperatorOAuthFlow(
+                    state=flow.state,
+                    server_id=server.id,
+                    operator_principal=operator_principal,
+                    created_at=now,
+                    expires_at=flow.expires_at,
+                    redirect_uri=flow.redirect_uri,
+                    code_verifier=flow.code_verifier,
+                    client_id=flow.client_info.client_id or "",
+                    client_secret=flow.client_info.client_secret,
+                    client_secret_expires_at=flow.client_info.client_secret_expires_at,
+                    token_endpoint_auth_method=flow.client_info.token_endpoint_auth_method,
+                    token_endpoint=flow.token_endpoint,
+                    resource=flow.resource,
+                    scope=flow.scope,
+                )
+            )
+        return McpOperatorAuthStartResponse(
+            server_id=server.id, authorization_url=flow.authorization_url, expires_at=flow.expires_at
+        )
+
+    async def complete_callback(self, *, state: str, code: str) -> McpOperatorAuthStatus:
+        with self._sessions.begin() as session:
+            if (row := session.get(McpOperatorOAuthFlow, state)) is None:
+                raise HTTPException(status_code=404, detail="OAuth flow not found or already used")
+            session.delete(row)
+            flow = OperatorOAuthFlowState.from_row(row)
+        now = datetime.datetime.now(datetime.UTC)
+        if flow.expires_at < now:
+            raise HTTPException(status_code=410, detail="OAuth flow expired; start connection again")
+        token = await _exchange_operator_oauth_code(flow, code)
+        token_expires_at = _token_expires_at(token, now)
+        with self._sessions.begin() as session:
+            existing = session.get(
+                McpOperatorOAuthAssociation, (flow.server_id, flow.operator_principal), with_for_update=True
+            )
+            if existing is None:
+                existing = McpOperatorOAuthAssociation(
+                    server_id=flow.server_id,
+                    operator_principal=flow.operator_principal,
+                    created_at=now,
+                    updated_at=now,
+                    client_id=flow.client_id,
+                    client_secret=flow.client_secret,
+                    client_secret_expires_at=flow.client_secret_expires_at,
+                    token_endpoint_auth_method=flow.token_endpoint_auth_method,
+                    token_endpoint=flow.token_endpoint,
+                    resource=flow.resource,
+                    access_token=token.access_token,
+                    refresh_token=token.refresh_token,
+                    token_type=token.token_type,
+                    scope=token.scope or flow.scope,
+                    token_expires_at=token_expires_at,
+                )
+                session.add(existing)
+            else:
+                existing.updated_at = now
+                existing.client_id = flow.client_id
+                existing.client_secret = flow.client_secret
+                existing.client_secret_expires_at = flow.client_secret_expires_at
+                existing.token_endpoint_auth_method = flow.token_endpoint_auth_method
+                existing.token_endpoint = flow.token_endpoint
+                existing.resource = flow.resource
+                existing.access_token = token.access_token
+                existing.refresh_token = token.refresh_token
+                existing.token_type = token.token_type
+                existing.scope = token.scope or flow.scope
+                existing.token_expires_at = token_expires_at
+            return _oauth_status_from_row(flow.server_id, flow.operator_principal, existing)
+
+    def disconnect(self, *, server_id: str, operator_principal: str) -> None:
+        with self._sessions.begin() as session:
+            row = session.get(McpOperatorOAuthAssociation, (server_id, operator_principal), with_for_update=True)
+            if row is not None:
+                session.delete(row)
+
+    async def access_token_for(self, *, server: McpServerEntry, operator_principal: str) -> str | None:
+        if not _operator_oauth_enabled(server):
+            return None
+        with self._sessions.begin() as session:
+            row = session.get(McpOperatorOAuthAssociation, (server.id, operator_principal), with_for_update=True)
+            if row is None:
+                return None
+            now = datetime.datetime.now(datetime.UTC)
+            if row.token_expires_at is None or row.token_expires_at > now + MCP_OPERATOR_AUTH_REFRESH_SKEW:
+                return row.access_token
+            if not row.refresh_token:
+                return None
+            snapshot = OperatorOAuthRefreshState.from_row(row)
+        refreshed = await _refresh_operator_oauth_token(snapshot)
+        token_expires_at = _token_expires_at(refreshed, datetime.datetime.now(datetime.UTC))
+        with self._sessions.begin() as session:
+            row = session.get(McpOperatorOAuthAssociation, (server.id, operator_principal), with_for_update=True)
+            if row is None:
+                return None
+            row.updated_at = datetime.datetime.now(datetime.UTC)
+            row.access_token = refreshed.access_token
+            row.refresh_token = refreshed.refresh_token or row.refresh_token
+            row.token_type = refreshed.token_type
+            row.scope = refreshed.scope or row.scope
+            row.token_expires_at = token_expires_at
+            return row.access_token
+
+
+def _oauth_store(request: Request) -> PostgresMcpOperatorOAuthStore:
+    store = request.app.state.mcp_operator_oauth_store
+    if store is None:
+        raise HTTPException(status_code=503, detail="MCP operator OAuth database is not configured")
+    return cast(PostgresMcpOperatorOAuthStore, store)
+
+
+def _maybe_oauth_store(request: Request) -> PostgresMcpOperatorOAuthStore | None:
+    return cast(PostgresMcpOperatorOAuthStore | None, request.app.state.mcp_operator_oauth_store)
+
+
+OAuthStoreDep = Annotated[PostgresMcpOperatorOAuthStore, Depends(_oauth_store)]
+
+
+def _operator_principal(request: Request) -> str:
+    return request.headers.get("x-authentik-username") or "operator"
+
+
+def _public_base_url(request: Request, settings: Settings) -> str:
+    if settings.public_base_url:
+        return settings.public_base_url.rstrip("/")
+    proto = (request.headers.get("x-forwarded-proto") or request.url.scheme).split(",", 1)[0].strip()
+    host = (
+        (request.headers.get("x-forwarded-host") or request.headers.get("host") or request.url.netloc)
+        .split(",", 1)[0]
+        .strip()
+    )
+    return f"{proto}://{host}"
+
+
+def _oauth_status_from_row(
+    server_id: str, operator_principal: str, row: McpOperatorOAuthAssociation | None
+) -> McpOperatorAuthStatus:
+    if row is None:
+        return McpOperatorAuthUnconnected(server_id=server_id, operator_principal=operator_principal)
+    return McpOperatorAuthConnected(
+        server_id=server_id,
+        operator_principal=operator_principal,
+        connected_at=row.created_at,
+        token_expires_at=row.token_expires_at,
+        scope=row.scope,
+    )
+
+
+def _token_expires_at(token: OAuthToken, now: datetime.datetime) -> datetime.datetime | None:
+    if token.expires_in is None:
+        return None
+    return now + datetime.timedelta(seconds=token.expires_in)
+
+
+def _metadata_request_headers() -> dict[str, str]:
+    return {MCP_PROTOCOL_VERSION: LATEST_PROTOCOL_VERSION}
+
+
+async def _discover_protected_resource(
+    client: httpx.AsyncClient, server_url: str, auth_probe: httpx.Response
+) -> ProtectedResourceMetadata | None:
+    metadata_url = extract_resource_metadata_from_www_auth(auth_probe)
+    for url in build_protected_resource_metadata_discovery_urls(metadata_url, server_url):
+        response = await client.get(url, headers=_metadata_request_headers())
+        if metadata := await handle_protected_resource_response(response):
+            return metadata
+    return None
+
+
+async def _discover_oauth_metadata(
+    client: httpx.AsyncClient, server_url: str, resource_metadata: ProtectedResourceMetadata | None
+) -> OAuthMetadata | None:
+    auth_server_url = str(resource_metadata.authorization_servers[0]) if resource_metadata else None
+    for url in build_oauth_authorization_server_metadata_discovery_urls(auth_server_url, server_url):
+        response = await client.get(url, headers=_metadata_request_headers())
+        ok, metadata = await handle_auth_metadata_response(response)
+        if metadata:
+            return metadata
+        if not ok:
+            break
+    return None
+
+
+def _resource_for_oauth(server_url: str, resource_metadata: ProtectedResourceMetadata | None) -> str | None:
+    if resource_metadata is None:
+        return None
+    requested = resource_url_from_server_url(server_url)
+    configured = str(resource_metadata.resource)
+    if check_resource_allowed(requested_resource=requested, configured_resource=configured):
+        return configured
+    return requested
+
+
+async def _register_oauth_client(
+    client: httpx.AsyncClient,
+    server_url: str,
+    oauth_metadata: OAuthMetadata | None,
+    client_metadata: OAuthClientMetadata,
+) -> OAuthClientInformationFull:
+    if oauth_metadata and oauth_metadata.registration_endpoint:
+        registration_url = str(oauth_metadata.registration_endpoint)
+    else:
+        registration_url = urljoin(_authorization_base_url(server_url), "/register")
+    response = await client.post(
+        registration_url,
+        json=client_metadata.model_dump(by_alias=True, mode="json", exclude_none=True),
+        headers={"Content-Type": "application/json"},
+    )
+    return await handle_registration_response(response)
+
+
+def _authorization_base_url(server_url: str) -> str:
+    parsed = httpx.URL(server_url)
+    return f"{parsed.scheme}://{parsed.host}{f':{parsed.port}' if parsed.port else ''}"
+
+
+class _ResolvedOAuthClient(BaseModel):
+    """The authorization-server metadata and registered client obtained by probing an MCP
+    server — everything `_build_operator_oauth_flow` needs to assemble the auth request."""
+
+    client_info: OAuthClientInformationFull
+    oauth_metadata: OAuthMetadata | None = None
+    resource_metadata: ProtectedResourceMetadata | None = None
+    scope: str | None = None
+
+
+async def _resolve_operator_oauth_client(
+    server: McpServerEntry, server_url: str, redirect_uri: str
+) -> _ResolvedOAuthClient:
+    """Probe the MCP server, discover its authorization-server metadata, and obtain a client
+    registration — a configured `static_client_id`, or Dynamic Client Registration. All the
+    network I/O of starting an operator OAuth flow lives here."""
+    assert server.operator_oauth is not None
+    try:
+        async with httpx.AsyncClient(follow_redirects=False, timeout=10.0) as client:
+            auth_probe = await client.get(server_url, headers=_metadata_request_headers())
+            resource_metadata = await _discover_protected_resource(client, server_url, auth_probe)
+            oauth_metadata = await _discover_oauth_metadata(client, server_url, resource_metadata)
+            scope = (
+                " ".join(server.operator_oauth.scopes)
+                if server.operator_oauth.scopes is not None
+                else get_client_metadata_scopes(
+                    extract_scope_from_www_auth(auth_probe), resource_metadata, oauth_metadata
+                )
+            )
+            client_metadata = OAuthClientMetadata(
+                client_name=server.operator_oauth.client_name,
+                redirect_uris=[redirect_uri],
+                grant_types=["authorization_code", "refresh_token"],
+                response_types=["code"],
+                scope=scope,
+            )
+            if server.operator_oauth.static_client_id:
+                client_info = OAuthClientInformationFull(
+                    client_id=server.operator_oauth.static_client_id,
+                    redirect_uris=client_metadata.redirect_uris,
+                    grant_types=client_metadata.grant_types,
+                    response_types=client_metadata.response_types,
+                    scope=client_metadata.scope,
+                    client_name=client_metadata.client_name,
+                )
+            else:
+                client_info = await _register_oauth_client(client, server_url, oauth_metadata, client_metadata)
+    except Exception as e:
+        raise HTTPException(status_code=502, detail=f"failed to start MCP OAuth flow for {server.id}: {e}") from e
+    if not client_info.client_id:
+        raise HTTPException(status_code=502, detail=f"MCP OAuth registration for {server.id} did not return client_id")
+    return _ResolvedOAuthClient(
+        client_info=client_info, oauth_metadata=oauth_metadata, resource_metadata=resource_metadata, scope=scope
+    )
+
+
+async def _build_operator_oauth_flow(server: McpServerEntry, public_base_url: str) -> _BuiltOperatorOAuthFlow:
+    assert server.operator_oauth is not None
+    # operator_oauth is meaningless for an in-process server (there's no remote authorization
+    # server to run DCR/metadata discovery against); bind server_url to a local for stable
+    # mypy narrowing across the awaits.
+    assert server.server_url is not None
+    server_url = server.server_url
+    redirect_uri = f"{public_base_url}{MCP_OPERATOR_AUTH_CALLBACK_PATH}"
+    resolved = await _resolve_operator_oauth_client(server, server_url, redirect_uri)
+    client_info = resolved.client_info
+    oauth_metadata = resolved.oauth_metadata
+    resource_metadata = resolved.resource_metadata
+    scope = resolved.scope
+    pkce = PKCEParameters.generate()
+    state = secrets.token_urlsafe(32)
+    if oauth_metadata and oauth_metadata.authorization_endpoint:
+        auth_endpoint = str(oauth_metadata.authorization_endpoint)
+    else:
+        auth_endpoint = urljoin(_authorization_base_url(server_url), "/authorize")
+    if oauth_metadata and oauth_metadata.token_endpoint:
+        token_endpoint = str(oauth_metadata.token_endpoint)
+    else:
+        token_endpoint = urljoin(_authorization_base_url(server_url), "/token")
+    resource = _resource_for_oauth(server_url, resource_metadata)
+    params = {
+        "response_type": "code",
+        "client_id": client_info.client_id,
+        "redirect_uri": redirect_uri,
+        "state": state,
+        "code_challenge": pkce.code_challenge,
+        "code_challenge_method": "S256",
+    }
+    if resource:
+        params["resource"] = resource
+    if scope:
+        params["scope"] = scope
+    return _BuiltOperatorOAuthFlow(
+        state=state,
+        authorization_url=f"{auth_endpoint}?{urlencode(params)}",
+        expires_at=datetime.datetime.now(datetime.UTC) + MCP_OPERATOR_AUTH_FLOW_TTL,
+        redirect_uri=redirect_uri,
+        code_verifier=pkce.code_verifier,
+        client_info=client_info,
+        token_endpoint=token_endpoint,
+        resource=resource,
+        scope=scope,
+    )
+
+
+def _token_request_auth(
+    data: dict[str, str], client: _OperatorOAuthTokenClient
+) -> tuple[dict[str, str], dict[str, str]]:
+    headers = {"Content-Type": "application/x-www-form-urlencoded"}
+    if client.token_endpoint_auth_method == "client_secret_basic" and client.client_secret:
+        encoded_id = quote(client.client_id, safe="")
+        encoded_secret = quote(client.client_secret, safe="")
+        credentials = base64.b64encode(f"{encoded_id}:{encoded_secret}".encode()).decode()
+        headers["Authorization"] = f"Basic {credentials}"
+    elif client.token_endpoint_auth_method == "client_secret_post" and client.client_secret:
+        data["client_secret"] = client.client_secret
+    return data, headers
+
+
+async def _exchange_operator_oauth_code(flow: OperatorOAuthFlowState, code: str) -> OAuthToken:
+    data: dict[str, str] = {
+        "grant_type": "authorization_code",
+        "code": code,
+        "redirect_uri": flow.redirect_uri,
+        "client_id": flow.client_id,
+        "code_verifier": flow.code_verifier,
+    }
+    if flow.resource:
+        data["resource"] = flow.resource
+    data, headers = _token_request_auth(data, flow)
+    async with httpx.AsyncClient(timeout=10.0) as client:
+        response = await client.post(flow.token_endpoint, data=data, headers=headers)
+    if response.status_code != 200:
+        raise HTTPException(status_code=502, detail=f"MCP OAuth token exchange failed: {response.status_code}")
+    try:
+        return await handle_token_response_scopes(response)
+    except ValidationError as e:
+        raise HTTPException(status_code=502, detail=f"MCP OAuth token response was invalid: {e}") from e
+
+
+async def _refresh_operator_oauth_token(association: OperatorOAuthRefreshState) -> OAuthToken:
+    refresh_token = association.refresh_token
+    if not refresh_token:
+        raise RuntimeError("MCP OAuth association has no refresh token; reconnect in the console")
+    data: dict[str, str] = {
+        "grant_type": "refresh_token",
+        "refresh_token": refresh_token,
+        "client_id": association.client_id,
+    }
+    if association.resource:
+        data["resource"] = association.resource
+    data, headers = _token_request_auth(data, association)
+    async with httpx.AsyncClient(timeout=10.0) as client:
+        response = await client.post(association.token_endpoint, data=data, headers=headers)
+    if response.status_code != 200:
+        raise RuntimeError(f"MCP OAuth token refresh failed: {response.status_code}")
+    try:
+        return await handle_token_response_scopes(response)
+    except ValidationError as e:
+        raise RuntimeError(f"MCP OAuth refresh response was invalid: {e}") from e
+
+
+# The callback page is served here (not from the SPA) because the OAuth provider redirects
+# straight to this backend endpoint, where the code→token exchange runs; the page's only job
+# is to report the outcome and nudge any open console tab. The markup lives in a sibling
+# .html file (loaded once at import) so it stays lintable rather than a Python blob;
+# `string.Template` `$` placeholders avoid colliding with the CSS/JS braces.
+# TODO: make this a SPA-style page instead of a backend-served .html template — have the
+# callback run the token exchange, then redirect to a frontend SPA route (e.g.
+# /mcp-auth-complete?status=…) that renders the outcome and posts the BroadcastChannel nudge,
+# so all markup lives in the frontend rather than this .html + Python pair.
+_CALLBACK_TEMPLATE = Template((Path(__file__).parent / "mcp_operator_auth_callback.html").read_text(encoding="utf-8"))
+
+
+def _oauth_callback_response(ok: bool, message: str, *, status_code: int = 200) -> HTMLResponse:
+    title = "MCP account connected" if ok else "MCP account connection failed"
+    return HTMLResponse(
+        status_code=status_code,
+        content=_CALLBACK_TEMPLATE.substitute(
+            title=html.escape(title), message=html.escape(message), payload="mcpAuthChanged" if ok else "mcpAuthFailed"
+        ),
+    )
+
+
+@router.get("/api/mcp/operator-auth")
+async def mcp_operator_auth_statuses(
+    request: Request, settings: SettingsDep, oauth_store: OAuthStoreDep
+) -> McpOperatorAuthStatusResponse:
+    return oauth_store.list_statuses(servers=_load_servers(settings), operator_principal=_operator_principal(request))
+
+
+@router.post("/api/mcp/operator-auth/{server_id}/start")
+async def start_mcp_operator_auth(
+    server_id: str, request: Request, csrf_protect: Csrf, settings: SettingsDep, oauth_store: OAuthStoreDep
+) -> McpOperatorAuthStartResponse:
+    await csrf_protect.validate_csrf(request)
+    server = _server_entry(settings, server_id)
+    return await oauth_store.start_flow(
+        server=server,
+        operator_principal=_operator_principal(request),
+        public_base_url=_public_base_url(request, settings),
+    )
+
+
+@router.delete("/api/mcp/operator-auth/{server_id}")
+async def disconnect_mcp_operator_auth(
+    server_id: str, request: Request, csrf_protect: Csrf, oauth_store: OAuthStoreDep
+) -> McpOperatorAuthUnconnected:
+    await csrf_protect.validate_csrf(request)
+    operator_principal = _operator_principal(request)
+    oauth_store.disconnect(server_id=server_id, operator_principal=operator_principal)
+    return McpOperatorAuthUnconnected(server_id=server_id, operator_principal=operator_principal)
+
+
+@router.get(MCP_OPERATOR_AUTH_CALLBACK_PATH)
+async def mcp_operator_auth_callback(
+    oauth_store: OAuthStoreDep, state: str | None = None, code: str | None = None, error: str | None = None
+) -> HTMLResponse:
+    if error:
+        return _oauth_callback_response(False, f"MCP OAuth authorization failed: {error}", status_code=400)
+    if not state or not code:
+        return _oauth_callback_response(False, "MCP OAuth callback is missing state or code.", status_code=400)
+    try:
+        status = await oauth_store.complete_callback(state=state, code=code)
+    except HTTPException as e:
+        detail = e.detail if isinstance(e.detail, str) else "MCP OAuth callback failed."
+        return _oauth_callback_response(False, detail, status_code=e.status_code)
+    return _oauth_callback_response(True, f"Connected {status.server_id} for {status.operator_principal}.")
