+++ haku/console/tools/BUILD.bazel
@@ -56,6 +56,7 @@ py_library(
     deps = [
         "//haku/console:deps",
         "//haku/console:mcp_approval",
+        "//haku/console:mcp_operator_oauth",
         "@pypi//fastapi",
         "@pypi//pydantic",
     ],
